fix postgres.d(44): Error: cannot implicitly convert expression (PQescapeString(buffer,cast(const(char*))cast(immutable(char)*)sqlData,sqlData.length)) of type ulong to int

+++ postgres.d
@@ -41,7 +41,7 @@ class PostgreSql : Database {
 
 	string escape(string sqlData) {
 		char* buffer = (new char[sqlData.length * 2 + 1]).ptr;
-		int size = PQescapeString (buffer, sqlData.ptr, sqlData.length);
+		ulong size = PQescapeString (buffer, sqlData.ptr, sqlData.length);
 
 		string ret = assumeUnique(buffer[0..size]);
